requirements.txt.txt
google-adk
fastapi
uvicorn
torch
transformers
peft
bitsandbytes
accelerate
huggingface_hub
datasets
pyngrok